Fixed movement issues. Fixed app launch defect. Changed media player.

+++ game-space-invaders/src/main/java/com/vcs/fx/game/model/Point.java
@@ -2,44 +2,44 @@ package com.vcs.fx.game.model;
 
 public class Point {
 
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     public Point(Point point) {
         this.x = point.x;
         this.y = point.y;
     }
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void increaseX(int x) {
+    public void increaseX(double x) {
         this.x += x;
     }
 
-    public void increaseY(int y) {
+    public void increaseY(double y) {
         this.y += y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public Point move(int yOffset) {
+    public Point move(double yOffset) {
         y += yOffset;
         return this;
     }
